// src/store/price-list/store/action-types.ts
export enum PriceListActionTypes {
    GET_PRICE_LIST_START = "GET_PRICE_LIST_START",
    GET_PRICE_LIST_SUCCESS = "GET_PRICE_LIST_SUCCESS",
    GET_PRICE_LIST_FAIL = "GET_PRICE_LIST_FAIL",

    GET_PRICE_LIST_BY_ID_START = "GET_PRICE_LIST_BY_ID_START",
    GET_PRICE_LIST_BY_ID_SUCCESS = "GET_PRICE_LIST_BY_ID_SUCCESS",
    GET_PRICE_LIST_BY_ID_FAIL = "GET_PRICE_LIST_BY_ID_FAIL",


    GET_PRICE_LIST_ALL_START = "GET_PRICE_LIST_ALL_START",
    GET_PRICE_LIST_ALL_SUCCESS = "GET_PRICE_LIST_ALL_SUCCESS",
    GET_PRICE_LIST_ALL_FAIL = "GET_PRICE_LIST_ALL_FAIL",
}